requirements.txt
   Flask==2.3.2
   Pillow==9.5.0
   PyPDF2==3.0.1
   pdf2image==1.16.3
   reportlab==3.6.12
   Werkzeug==2.3.3
   google-cloud-vision==3.4.1
   fpdf2